% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{build_solr_query}
\alias{build_solr_query}
\title{Solr query}
\usage{
build_solr_query(...)
}
\arguments{
\item{...}{Named arguments to build Solr query from}
}
\value{
String with Solr query
}
\description{
Build Solr query from named arguments
}
\keyword{internal}
