% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{call_mc_api}
\alias{call_mc_api}
\title{Call API}
\usage{
call_mc_api(url)
}
\arguments{
\item{url}{A URL string built with create_mc_url()}
}
\value{
JSON content object
}
\description{
Call API, check for status code and return content
}
