% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{check_rate_limit}
\alias{check_rate_limit}
\title{Check rate limit}
\usage{
check_rate_limit(key = NULL)
}
\arguments{
\item{key}{MediaCloud API key. Will be read from environment
variable 'MEDIACLOUD_API_KEY' if set to \code{NULL} (default).}
}
\value{
A tibble with information about the current rate limits.
}
\description{
Check your current rate limit
}
