% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{create_mc_url}
\alias{create_mc_url}
\title{Construct Request URL}
\usage{
create_mc_url(
  endpoint,
  path = NULL,
  parameters = NULL,
  key = Sys.getenv("MEDIACLOUD_API_KEY")
)
}
\description{
Construct Request URL
}
