% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\docType{package}
\name{mediacloud}
\alias{mediacloud}
\title{mediacloud: R wrapper package for the MediaCloud API}
\description{
A simple API wrapper for the MediaCloud API. Curenntly supports
searching for media outlets and stories, getting word matrices for stories,
getting single stories, and checking rate limits. Calls to the stories and
word matrices endpoints have simplified interface for constructing the
necessary Solr queries. All results follow tidy data principles.
}
