% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{search_stories}
\alias{search_stories}
\title{Search stories}
\usage{
search_stories(
  text = NULL,
  media_id = NULL,
  stories_id = NULL,
  publish_year = NULL,
  publish_month = NULL,
  publish_day = NULL,
  publish_date = NULL,
  n = 20,
  last_processed_stories_id = NULL,
  key = NULL,
  tibble = TRUE
)
}
\description{
Search stories
}
