% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{get_word_matrices}
\alias{get_word_matrices}
\title{Get word matrices}
\usage{
get_word_matrices(
  text = NULL,
  media_id = NULL,
  stories_id = NULL,
  publish_year = NULL,
  publish_month = NULL,
  publish_day = NULL,
  publish_date = NULL,
  n = 20,
  stopword_length = NULL,
  key = NULL,
  tibble = TRUE
)
}
\description{
Get word matrices
}
